Fix warning on some builds
PiperOrigin-RevId: 893196389

+++ test/subgraph/input-output.cc
@@ -174,7 +174,7 @@ TEST_P(InputOutput, SlidingWindow) {
   // We should start out with a buffer of zeros.
   std::vector<float> expected(dims.front(), 0.0f);
   for (size_t t = 0; t < slices * 2; ++t) {
-    std::fill(input.begin(), input.end(), t);
+    std::fill(input.begin(), input.end(), static_cast<float>(t));
     subgraph.InvokeRuntime();
 
     for (size_t i = 0; i < slices; ++i) {
